refactor: make library warning-clean under -Wall -Wextra
Fix all 67 -Wall -Wextra warnings in src/ without changing behavior, and
replace the non-standard <memory.h> with <string.h> throughout:

- sign-compare: cast non-negative int counters at unsigned comparison sites
- maybe-uninitialized: initialize locals to neutral defaults
- pointer-sign: cast unsigned char* <-> char* at the boundary (IBMP/IGIF/IXPM)
- unused-parameter: explicit (void) for unused 'options' params
- unused-result: check fgets/fread returns in IPPM, returning IFileInvalid
- absolute-value: use fabs() for doubles in IDrawLin (also fixes a latent dash
  bug where abs() truncated the slope to int)
- unused-but-set / type-limits / int-to-pointer: remove dead vars; drop the
  always-true unsigned >= 0 test; cast via intptr_t in IColor

IWriteImageFile now returns IFunctionNotImplemented for formats without a
write function instead of an uninitialized value (fixes both the warning and
a real garbage-return bug).

+++ src/ICopyImg.c
@@ -22,7 +22,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <ctype.h>
-#include <memory.h>
+#include <string.h>
 
 #include "Ilib.h"
 #include "IlibP.h"
@@ -72,8 +72,8 @@ int dest_x, dest_y;
   colorp = _IGetColor ( color );
   ISetForeground ( gc, color );
 
-  for ( row = src_y, y = dest_y; row < src_y + height; row++, y++ ) {
-    for ( col = src_x, x = dest_x; col < src_x + width; col++, x++ ) {
+  for ( row = src_y, y = dest_y; row < src_y + (int) height; row++, y++ ) {
+    for ( col = src_x, x = dest_x; col < src_x + (int) width; col++, x++ ) {
       if ( i1->greyscale ) {
         ptr = i1->data + ( row * i1->width ) + col;
         colorp->red = colorp->green = colorp->blue = *(ptr);
@@ -164,8 +164,8 @@ unsigned int dest_width, dest_height;
   */
   scalex = (double) dest_width / (double) src_width;
   scaley = (double) dest_height / (double) src_height;
-  for ( y = dest_y; y < dest_y + dest_height; y++ ) {
-    for ( x = dest_x; x < dest_x + dest_width; x++ ) {
+  for ( y = dest_y; y < dest_y + (int) dest_height; y++ ) {
+    for ( x = dest_x; x < dest_x + (int) dest_width; x++ ) {
       /* get location from source image for this location
       ** x2,y2 is location in source image.
       */
